feat(sandbox): Support running library benchmarks in a sandbox
